refactor : IDE-117 파일 열기 함수 커스텀 hook 분리용
